Removed passing storageEngine from globalInitializer. Added protected route to dashboards. Moved effect for logging in user to Login
