+++ src/GlobalInitializer.tsx
@@ -16,6 +16,6 @@ export function GlobalInitializer() {
     fn();
   }, [setStorageEngine, storageEngine]);
   return (
-    <GlobalConfigParser storageEngine={storageEngine} />
+    <GlobalConfigParser />
   );
 }
